requirements.txt
pandas==2.1.4
geopandas==0.14.2
keplergl==0.3.2
pillow==10.2.0
matplotlib==3.8.2
scikit-learn==1.3.2
torch==2.1.2
torchvision==0.16.2
numpy==1.26.3
tqdm==0.4.6
h3==3.7.6
h3pandas==0.2.6
umap-learn==0.5.1
umap==0.1.1
umap-learn==0.5.1